+++ src/components/datagrid/data_grid.tsx
@@ -29,7 +29,7 @@ import {
   EuiDataGridStyleRowHover,
   EuiDataGridPopoverContents,
   EuiDataGridColumnVisibility,
-  EuiDataGridTooBarVisibilityOptions,
+  EuiDataGridToolBarVisibilityOptions,
 } from './data_grid_types';
 import { EuiDataGridCellProps } from './data_grid_cell';
 // @ts-ignore-next-line
@@ -90,9 +90,9 @@ type CommonGridProps = CommonProps &
      */
     gridStyle?: EuiDataGridStyle;
     /**
-     * Accepts either a boolean or #EuiDataGridToolbarVisibilityOptions object. When used as a boolean, defines the display of the toolbar entire. WHen passed an object allows you to turn off individual controls within the toolbar.
+     * Accepts either a boolean or #EuiDataGridToolbarVisibilityOptions object. When used as a boolean, defines the display of the toolbar entire. WHen passed an object allows you to turn off individual controls within the toolbar as well as add additional buttons.
      */
-    toolbarVisibility?: boolean | EuiDataGridTooBarVisibilityOptions;
+    toolbarVisibility?: boolean | EuiDataGridToolBarVisibilityOptions;
     /**
      * A #EuiDataGridInMemory object to definite the level of high order schema-detection and sorting logic to use on your data. *Try to set when possible*. When ommited, disables all enhancements and assumes content is flat strings.
      */
@@ -552,20 +552,26 @@ export const EuiDataGrid: FunctionComponent<EuiDataGridProps> = props => {
   // By default the toolbar appears
   const showToolbar = !!toolbarVisibility;
 
-  // Typegaurd to see if toolbarVisibility has a certain boolean property assigned
+  // Typeguards to see if toolbarVisibility has a certain boolean property assigned
   // If not, just set it to true and assume it's OK to show
-  function checkOrDefaultToolBarDiplayOptions(
+  function objectHasKey<
+    O extends Record<string, any>,
+    ObjectKey extends keyof O
+  >(object: O, key: ObjectKey): object is Required<O> {
+    return object.hasOwnProperty(key);
+  }
+  function checkOrDefaultToolBarDiplayOptions<
+    OptionKey extends keyof EuiDataGridToolBarVisibilityOptions
+  >(
     arg: EuiDataGridProps['toolbarVisibility'],
-    option: keyof EuiDataGridTooBarVisibilityOptions
-  ): boolean {
+    option: OptionKey
+  ): Required<EuiDataGridToolBarVisibilityOptions>[OptionKey] {
     if (arg === undefined) {
       return true;
     } else if (typeof arg === 'boolean') {
       return arg as boolean;
-    } else if (
-      (arg as EuiDataGridTooBarVisibilityOptions).hasOwnProperty(option)
-    ) {
-      return arg[option]!;
+    } else if (objectHasKey(arg, option)) {
+      return arg[option];
     } else {
       return true;
     }
@@ -589,6 +595,12 @@ export const EuiDataGrid: FunctionComponent<EuiDataGridProps> = props => {
         : null}
       {checkOrDefaultToolBarDiplayOptions(toolbarVisibility, 'showSortSelector')
         ? columnSorting
+        : null}
+      {checkOrDefaultToolBarDiplayOptions(
+        toolbarVisibility,
+        'additionalControls'
+      ) && typeof toolbarVisibility !== 'boolean'
+        ? toolbarVisibility.additionalControls
         : null}
     </Fragment>
   );