+++ src/components/datagrid/data_grid_types.ts
@@ -67,7 +67,7 @@ export interface EuiDataGridStyle {
   cellPadding?: EuiDataGridStyleCellPaddings;
 }
 
-export interface EuiDataGridTooBarVisibilityOptions {
+export interface EuiDataGridToolBarVisibilityOptions {
   /**
    * Allows the ability for the user to hide fields and sort columns
    */
@@ -84,6 +84,10 @@ export interface EuiDataGridTooBarVisibilityOptions {
    * Allows user to be able to full screen the data grid. If set to `false` make sure your grid fits within a large enough panel to still show the other controls.
    */
   showFullScreenSelector?: boolean;
+  /**
+   * Will place any passed node into the toolbar in front of the fullscreen button. Recommend using EuiButtonEmpty with the props shown in the examples.
+   */
+  additionalControls?: ReactNode;
 }
 
 // ideally this would use a generic to enforce `pageSize` exists in `pageSizeOptions`,