+++ src/components/search_bar/index.js
@@ -1,3 +1,6 @@
 export { EuiSearchBar, QueryType, Query, Ast } from './search_bar';
 export { SearchBoxConfigPropTypes } from './search_box';
 export { SearchFiltersFiltersType } from './search_filters';
+
+// TODO: Some related types are defined in basic_table/in_memory_table.
+// Use and remove them when TypeScriptification is done.